requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
dacite==1.9.2
Deprecated==1.2.18
docutils==0.22
fastapi==0.103.2
filetype==1.2.0
Flask==3.1.2
Flask-WTF==1.2.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.6
Kivy==2.3.1
Kivy-examples==2.3.1
Kivy-Garden==0.1.5
macholib==1.16.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mutagen==1.47.0
mysql-connector-python==9.4.0
numpy==2.3.3
packaging==25.0
pillow==10.4.0
platformdirs==4.4.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pykakasi==2.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytube==15.0.0
RapidFuzz==3.14.1
redis==6.4.0
reportlab==4.4.4
requests==2.32.5
rich==13.9.4
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundcloud-v2==1.6.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
spotdl==4.4.2
spotipy==2.25.1
starlette==0.27.0
syncedlyrics==1.0.1
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.23.2
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
yt-dlp==2025.9.5
ytmusicapi==1.11.1
